Added grappling hook, tweaked player animations, switched joystick assets, camera tweaks,
Added basic grappling hook logic. Not tested with multiplayer yet. The hook isn't even synced, but the player position should work.

Changed player character to the Kenney alien instead of a spaceship. Tweaked the player animations so we see the character standing while immobile on ground, they have a "jump" animation when in the air, and a "walk" animation while moving on ground.

The old joystick assets were dark and didn't look good colorized

Zoomed the camera out from previous to see more of the playing field. On clients, made the camera follow the local player. No solution for server yet, but probably some kind of automatic zooming to show all players?