Added a boundary around the level, and added camera clamping to the bounds
To avoid players flying out of the level, added some boundaries around it. It's not quite perfect, as the corners don't have super accessible grappling points, but at least you can use the floor if needed.

The camera gets clamped to stay within the map. Tested this just with the client, but it should work on the server as well.