Added syncing of grappling hook state across network, moved more authority to server
Instead of a couple of booleans, I'm now using an enum to track the hook's state so it can be sent over the network as a single variable.

As I've moved more authority to the server, in the current version it's not actually possible to use the hook or shoot projectiles without a server present. I tried simply changing the "call_remote" RPC's to "call_local" but that wan't sufficient. I will investigate this further later. Hopefully I don't have to make too much local-specific logic to make it work. The server being in charge is certainly not a bad thing for multiplayer purposes, but can be a bit cumbersome for testing.