Fixed the camera clamping so it handles weird aspect ratios correctly
Previously it wouldn't quite work right if the aspect ratio wasn't 16:9. Now it can deal with any aspect ratio, though in extreme cases the map will be in the top edge or the left edge of the screen. That seems like an unnecessary edge case to deal with, though for very small maps it could be a minor problem.