Added some basic explosion forces
Added a basic velocity modification for when a player character is within an explosion radius.

I think I need to make the explosions much bigger, and I also think the force needs to be normalized so that a direct hit will always apply maximum force. Right now it just scales up the closer the character is to the middle of the explosion.

Also, as these explosions have sent the character flying kind of fast, I'm realizing my previous change to the projectile starting velocities will probably need to be reversed, as I had removed the compound effect of the character's velocity as I figured it would be more consistent for network syncing. It doesn't feel right though, so that's a problem to sovle.