Added tracking of projectiles by ID
The ID is determined by the client, but is specific to each player/client, so there should be no overlaps. The ID is used for removing the projectile from the scene once it hits a target. Previously this would only happen on server side as the handling of the collision was only handled on server side.