Added measuring of network latency
The clients will ping the server ten times a second (adjustable) and calculate a rolling one-second average of those. This is displayed on screen for now as I want to have easy access to that information during testing, though I will likely eventually remove it. I intend to use the latency measurement for extrapolation of object state, e.g. if I know a player's location and velocity from 100ms ago, I can calculate where they should be now, and show the player a more accurate game state on the client side.