Added an estimation of the current server time on client
Tried a couple of approaches, but for now settled on a rolling average of the estimated difference between local time and server time, the server time calculated from a timestamp sent by the server plus the network latency. Later I'll add timestamps to other network events, which can then be used to extrapolate game state to something closer to the current server-side state.