Added a simple "main menu"
The menu isn't very functional yet, but I want to add a way to play the game without a remote server, so this menu will have the option of either joining a remote server, or hosting your own game. Right now both buttons just load the main scene, and the "server address" text box is placeholder.