Split most of the logic from main.gd to server.gd and client.gd
It hadn't been clear to me how to handle RPC across different scripts, but it seems like the way to do it is to give the two nodes the same name, and have both scripts include the same RPC functions with the same signatures.

This has produced some duplicate code, as both the client and server are running a fair bit of the same logic. I will likely extract some of the logic into shared functions, but I wonder if I may need to go back to running the same script on client and server, as I want it to be possible to run the game without a dedicated server with one of the players hosting the game.

As a part of this process, I also extracted the tilemap level into its own scene. This was previously a part of the Main scene, but now both Client and Server scenes include the Level scene. I also duplicated the Camera2D node to both Server and Client.