Main menu now loads either server or client from the two buttons
Made the main menu load the server scene or the client scene depending on which button the player presses. This exposed an issue where I was still relying on my custom "server" feature tag to choose which camera mode to use, so I added a boolean variable for it instead, which is now different for the camera script in the Client and Server scenes, so each uses the right camera mode.