Moved some of the grappling hook init code into the grappling hook script
This turned out to be even less code than with the projectile init, but I figure it's probably still worth it, and it makes more sense for the logic to be a part of the grappling hook script anyway. In addition to the actual code, I also moved the "hooks" dictionary to the grappling hook script as a static variable.