Extracted shared code from server and client script to a shared parent class called Sync
In retrospect this feels very obvious, but I'm not used to GDScript and it took a while for inheritance to even occur to me. I think this feels like a nice way to share the code, as the identical parts can be omitted completely from the child scripts, so this makes it much easier to see the more relevant stuff.

One wrinkle I did find is in the sync_projectile_shot RPC function, as most of the logic is shared between server and client, but on the server side I want to connect some signal listeners, and it turns out I can't return a value from the RPC function, so I had to make a separate function that creates the projectile scene, which I call from the Sync class (used by the Client), and on the Server I call it and use the return value to connect the signals. Not too big of a deal, but it would have been cleaner if I didn't need to wrap that logic in a separate function.