Fixed a bug (?) with the ping code
I was sending the ping to multiplayer ID 0 instead of ID 1, which is reserved for the server. This was working fine though, and I'm not sure why. My best guess is that ID 0 meant that the ping got sent to all network peers, but since the clients don't respond to pings this didn't cause any trouble.