Added passing the server address from the main menu to the client, and saving it in a file
Connected the text box in the main menu so the entered address is passed to the client script. Also I added saving the server address into a file, because I wanted to be able to connect to a remote server without typing the address in every time and without including it in the game data.