Added different player character colors, chosen based on network ID
I wanted the players to be more easily identifiable, and since the Kenney spritesheet I'm using has five different colored versions of the character, I added these as separate resources, and I choose one based on the player's network ID. This way I don't need to sync the information about the player's color but it'll be still consistent across different clients. The downside of course is that it is still possible to end up with two players with the same color by chance. To reduce the chance, I added all five colors so it's at least less likely.