Improved the spectator camera a bit
I had used the average of the players' positions, but with more than two players I actually want the midpoint of the bounding box that contains all the players.

I am finding that with this new logic the zoom does sometimes push the camera out of the level bounds, so I still need to revise this, but at least it now zooms out enough to show all players.