Added basic scoring
Clients show a scoreboard with each player's portrait (color) shown next to their score. I made this choice but actually already in testing I realized I probably want to add it on the server display as well, so I may move some of the relevant code to sync.gd from client.gd later.

The scoring is tracked on the server and synced to clients in the sync_player_state RPC. When a player dies, the killer gets a point, unless the killer killed themselves, in which case they lose a point. To support this, I added a `player` field to the Explosion so we know who originated the explosion. This info did already exist in the projectile script, which is where we get it from.