Added explicit "player_join_game" RPC, and a "sync_player_list" RPC
This is in preparation for supporting hosting a game without  a dedicated server. Incidentally also changed the code slightly so that the players dictionary also contains the local player, and there's an array called player_ids which in part is used to sort the scoreboard, as it's stored in the order received from the server via sync_player_list.