Added support for the host to have a local player
Essentially I replicated logic from server.gd into client.gd, and added some specific code to make sure that the local player is handled correctly.

This does duplicate some code in server.gd and client.gd, but I may eventually just get rid of server.gd. For now, I did change the main scene of the project to the Main scene, and updated it so it either spawns the main menu scene, or if it's a dedicated server build, it spawns the server scene. That is still using the server script. The downside with that one is that it doesn't yet have the scoreboard, so I think I will set up the client script with an option to get rid of the local player, and then I'll just use the same scene and script for dedicated server.