Enabled dedicated server to use the same "client" scene and script
Added logic to the client script so it gets rid of the local player node and some of the player UI (touch joysticks and the latency indicator) if the build has the "server" feature, which I use for the dedicated server builds. Completely got rid of the server script and scene, as they are no longer needed. Changed the main script to load the Client scene instead of the Server scene for dedicated server use, and added the is_host flag there so the script will behave like a server.