Network sync tweaks
Added some basic accounting for network latency. Added timestamps to sync_player_state RPCs and if the state is more than one frame old, run the simulation extra times to catch up.

For projectiles and the grappling hook, changed the protocol so that instead of sending the spawn position, direction and speed, all that gets sent is a target position. This is analogous to the direction, but should be more resiliant against latency, since the shot will be fired toward the same target regardless of how far the player character will have moved since the message was sent. The main purpose of this change is to prevent cheating through a hacked client, as the client is unable to send fraudulent data, e.g. spawning projectiles where they shouldn't be able to.

Also added some extra tags to the RPC function definitions: "reliable" for many of the functions, as most major events should not be missed. For sync_player_state I'm using "unreliable_ordered" which means old packets will be thrown away if they arrive out of order. I feel this is better particularly now that I'm doing the catchup frames.

Thinking about whether I should add syncing of projectile states. It's not strictly necessary, since projectiles only do things on server side, but maybe it would be beneficial to at least do an initial sync so that clients have accurate information on where the projectile was spawned from. Since the trajectories are completely predictable, a single sync like that is probably sufficient.