Made the server only send the full player state at given intervals
Now the server only sends the sync_player_state message ten times a second. It's not strictly necessary, but I wanted to see if that's viable, and it seems to work just fine, probably in large part because the simulation is fairly predictable even without the full syncs.

In addition to sending the update on an interval, I also send it if the player has just died or just respawned, since there are no specific events for that.