Fixed a bug in the ping logic
In the ping logic, I had added a variable _waiting_for_ping to make sure we don't send a new ping before receiving the previous one... and then I forgot to actually set the variable, so the system would just keep sending more pings, and with high enough latency this was messing with the latency measurements since I have not set it up to deal with multiple simultaneous requests.

In related news, I started doing some testing with a network latency simulation tool, and it definitely looks like I have some work to do there. First priority are the projectiles.