Changed projectile spawning logic, so clients don't spawn projectiles until the server tells them to
sync_projectile_shot is now only handled by the server, and the server then sends a sync_projectile_state RPC to the clients. Also extracted the network latency catchup code into its own function so I can do this for projectiles as well. The projectile state is only synced when the projectile is spawned, since the trajectory should be completely predictable.