Added removing projectiles locally on impact.
When an impact is detected for a projectile clientside, the projectile is now removed from the scene. Otherwise it just looks buggy if there's high latency. The explosion isn't spawned at this point, though maybe I should consider that at least for stationary targets and the projectile reaching its maximum range.