Added a delay for the local player in spawning their grappling hook and projectiles in an effort to spawn it in sync with the server
Changed the projectile shot logic so the sync_projectile_shot event is sent to all players instead of just the server. For other clients, they'll just spawn the projectile immediately, which should be ok assuming all players have similar network latency. That's not universally true but probably good enough, at least for now. The local player however will purposefully delay the spawning of their projectiles to try to spawn the projectiles in sync with the server. Also added similar logic for the grappling hook. Moved sync_grapplinghook_shot into client.gd from sync.gd since the latency data is stored in client.gd .

I've been thinking maybe it would be a good idea to build some latency into the actions by default, so that it will feel the same regardless of how much network latency there is. I'm not sure how this could be done without opening any possibility for cheating though, since the client would need to tell the server when the projectile is supposed to be spawned, but it feels like it would be in the client's benefit to tell the server the projectile should be spawned as early as possible.