Added a bit of code to update the grappling hook's visuals in _ready
I had noticed the grappling hook looking kind of buggy sometimes when it spawns, which is likely since the default positions for the line renderer's points would be in some random location. Added a bit of code that updates the points in _ready so they should be at least somewhat reasonable when created.