Fixed some bugs with the grappling hook, and an issue where a client player could see themselves in the wrong location for a moment while waiting for the server to update their position at the start of the game
In particular, I realized the hooks weren't really getting cleaned up properly after reeling, as the signal that was emitting wasn't actually connected to anything. It would have been a slight hassle to connect it though, so for now I just added a call to detach_hook in that part of the hook's update, and this seems to work fine. Also, as I was figuring out what the problem with this was, I also added a bit of logic that triggers a player state update if the grappling hook has changed state. I do wonder if it might be better to make the grappling hook update its own RPC, but the player's state will of course start changing when the grappling hook attaches, so that's not a bad moment to update their state.

Also I noticed while testing with high simulated network latency that when a client joins the game, there is a moment before the server has updated their location, and the player can appear in the wrong location for a moment, so I fixed this by setting the player's health to 0 at the start of the local_player script. This will be automatically updated to 100 once the client receives the first state update from the server. For a host player, added a bit of code to update the health to 100 at the start.