Camera smoothing and look-ahead
Added some basic smoothing to the camera using lerp; in particular to get rid of the jarring snap that would happen when the player respawns or dies and the camera switches between player mode and spectator mode, or on the spectator camera when a player suddenly appears or disappears from existence. The smoothing is applied both to zooming and camera position.

Made the player camera look ahead of the player by one second, and made the spectator camera take that motion into account when determining the bounding box around the players, so if a player is fast moving toward the edge of the bounding box, the bounding box expands so it doesn't lose track of the players even with the smoothing applied.