Updated the touch joystick code to produce Godot input events
The previous version would read the mouse or touch inputs and store that info in the touch joystick node script's variables, which were then read by the gameplay code. It would also itself read input actions and translate that into "touch joystick inputs" as well so the game could optionally be played with a gamepad. Now I've removed that translation, and the touch inputs are instead translated into Godot input actions using the Input.parse_input_event function. The gameplay code then just reads the input actions as if they were coming from a regular gamepad, and this also works with an actual gamepad.

This should make it a fair bit simpler to implement player inputs in future projects, and also removes the dependency between the gameplay code and the touch joystick code, since the gameplay code just cares about Godot input actions. I think this approach would also make it relatively easy to implement touch controls into a game designed for a gamepad.

I already had the input action names set up in the script as I was doing the translation in the other direction before.